requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.2.0
Deprecated==1.2.13
Django==4.0.2
django-celery-results==2.2.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
kombu==5.2.3
konfig==1.1
mysqlclient==2.1.0
packaging==21.3
prompt-toolkit==3.0.28
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
